requirements.txt
fastapi
uvicorn
python-dotenv
motor
pydantic
pymongo
python-jose[cryptography]
passlib==1.7.4
bcrypt==3.2.2
email-validator
fastapi-mail
